TDM: fix some compile warnings
This is to avoid the below warnings:

drivers/tdm/device/tdm_fsl.c: In function 'tdm_fsl_reg_init':
drivers/tdm/device/tdm_fsl.c:416:14: warning:
unused variable 'base' [-Wunused-variable]
   phys_addr_t base = get_immrbase();
               ^

drivers/tdm/device/tdm_fsl.c:202:20: warning: 'tdm_err_isr'
defined but not used [-Wunused-function]
 static irqreturn_t tdm_err_isr(int irq, void *p)
                    ^

drivers/tdm/test/tdm_test.c:59:25: warning: 'tdmport'
defined but not used [-Wunused-variable]
 static struct tdm_port *tdmport;
				^

drivers/tdm/tdm-core.c: In function 'tdm_device_match':
drivers/tdm/tdm-core.c:74:11: warning: cast from pointer to
integer of different size [-Wpointer-to-int-cast]
    return (int)driver->id_table;
           ^

drivers/tdm/line_ctrl/slic_zarlink.c: In function 'configure_spi_pdata':
drivers/tdm/line_ctrl/slic_zarlink.c:573:22: warning:
 unused variable 'np' [-Wunused-variable]
  struct device_node *np = spi->dev.of_node;
                      ^
drivers/tdm/line_ctrl/slic_zarlink.c:572:13: warning:
 unused variable 'iprop' [-Wunused-variable]
  const u32 *iprop;
             ^

+++ drivers/tdm/tdm-core.c
@@ -66,12 +66,13 @@ static int use_latest_tdm_data = 1;
 static void tdm_data_tasklet_fn(unsigned long);
 
 /* tries to match client driver with the adapter */
-static int tdm_device_match(struct tdm_driver *driver, struct tdm_adapter *adap)
+static long tdm_device_match(struct tdm_driver *driver,
+				struct tdm_adapter *adap)
 {
 	/* match on an id table if there is one */
 	if (driver->id_table && driver->id_table->name[0]) {
 		if (!(strcmp(driver->id_table->name, adap->name)))
-			return (int)driver->id_table;
+			return (long)driver->id_table;
 	}
 	return TDM_E_OK;
 }
